LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-base-element/base-javascript-expected.txt

CONSOLE MESSAGE: Blocked setting javascript:/,test as the base URL because it does not have an allowed scheme.
CONSOLE MESSAGE: Blocked setting javascript:/,test as the base URL because it does not have an allowed scheme.
CONSOLE MESSAGE: Blocked setting javascript:/,more-test as the base URL because it does not have an allowed scheme.

PASS First <base> has a javascript: URL so fallback is used
PASS First <base> is removed so second is used
PASS Dynamically inserted first <base> has a javascript: URL so fallback is used
